Tests

+++ packages/core/src/fields/types/file/tests/test-fixtures.ts
@@ -2,8 +2,10 @@ import path from 'path';
 import fs from 'fs-extra';
 import { Upload } from 'graphql-upload';
 import mime from 'mime';
+import fetch from 'node-fetch';
 import { file } from '..';
 import { expectSingleResolverError } from '../../../../../../../tests/api-tests/utils';
+import { KeystoneConfig } from '../../../../types/config';
 
 const prepareFile = (_filePath: string) => {
   const filePath = path.resolve(`${__dirname}/../test-files/${_filePath}`);
@@ -16,6 +18,33 @@ const prepareFile = (_filePath: string) => {
     encoding: 'utf-8',
   });
   return { upload };
+};
+
+export const testMatrix = ['local'];
+
+if (process.env.S3_BUCKET_NAME) {
+  testMatrix.push('s3');
+}
+
+export const getRootConfig = (matrixValue: 's3' | 'local'): Partial<KeystoneConfig> => {
+  if (matrixValue === 'local') {
+    return {};
+  }
+  return {
+    files: {
+      upload: 's3',
+    },
+    experimental: {
+      s3: {
+        bucketName: process.env.S3_BUCKET_NAME!,
+        accessKeyId: process.env.S3_ACCESS_KEY_ID!,
+        secretAccessKey: process.env.S3_SECRET_ACCESS_KEY!,
+        region: process.env.S3_REGION!,
+        endpoint: process.env.S3_ENDPOINT,
+        forcePathStyle: process.env.S3_FORCE_PATH_STYLE === 'true',
+      },
+    },
+  };
 };
 
 export const name = 'File';
@@ -34,9 +63,11 @@ export const subfieldName = 'filesize';
 
 export const getTestFields = () => ({ secretFile: file() });
 
-export const afterEach = async () => {
-  // This matches the storagePath in the keystone config in the various test files.
-  fs.rmdirSync('tmp_test_files', { recursive: true });
+export const afterEach = async (matrixValue: 's3' | 'local') => {
+  if (matrixValue === 'local') {
+    // This matches the storagePath in the keystone config in the various test files.
+    fs.rmdirSync('tmp_test_files', { recursive: true });
+  }
 };
 
 export const initItems = () => [
@@ -65,11 +96,12 @@ export const crudTests = (keystoneTestWrapper: any) => {
   describe('Create - upload', () => {
     test(
       'upload values should match expected',
-      keystoneTestWrapper(async ({ context }: { context: any }) => {
-        const filename = 'keystone.jpeg';
-        const data = await context.query.Test.createOne({
-          data: { secretFile: prepareFile(filename) },
-          query: `
+      keystoneTestWrapper(
+        async ({ context, matrixValue }: { context: any; matrixValue: 's3' | 'local' }) => {
+          const filename = 'keystone.jpeg';
+          const data = await context.query.Test.createOne({
+            data: { secretFile: prepareFile(filename) },
+            query: `
               secretFile {
                 filename
                 __typename
@@ -78,13 +110,28 @@ export const crudTests = (keystoneTestWrapper: any) => {
                 url
               }
           `,
-        });
-        expect(data).not.toBe(null);
-        expect(data.secretFile.ref).toEqual(`local:file:${data.secretFile.filename}`);
-        expect(data.secretFile.url).toEqual(`/files/${data.secretFile.filename}`);
-        expect(data.secretFile.filesize).toEqual(3250);
-        expect(data.secretFile.__typename).toEqual('LocalFileFieldOutput');
-      })
+          });
+          expect(data).not.toBe(null);
+          expect(data.secretFile.ref).toEqual(`${matrixValue}:file:${data.secretFile.filename}`);
+          // since it would be hard to assert exactly on the returned url for s3, we're gonna check the content instead.
+          if (matrixValue === 's3') {
+            expect(data.secretFile.url).toEqual(
+              expect.stringContaining(`/${data.secretFile.filename}`)
+            );
+            const contentFromURL = await fetch(data.secretFile.url).then(x => x.buffer());
+            const contentFromFile = await fs.readFile(
+              path.resolve(`${__dirname}/../test-files/${filename}`)
+            );
+            expect(contentFromURL).toEqual(contentFromFile);
+          } else {
+            expect(data.secretFile.url).toEqual(`/files/${data.secretFile.filename}`);
+          }
+          expect(data.secretFile.filesize).toEqual(3250);
+          expect(data.secretFile.__typename).toEqual(
+            matrixValue === 'local' ? 'LocalFileFieldOutput' : 'S3FileFieldOutput'
+          );
+        }
+      )
     );
   });
   describe('Create - ref', () => {
